create context text transform and fix bugs, warnigs
